Rename client/source.rs => grpc.rs
